Extracted segmentation utils function
